refactor(chat): extract streamText probe/retry loop into streamTextWithRecovery
